[Backend][US-016] write_mlir support spec/impl/miter: refactor string => string-util in infrastructure/utils
