+++ infrastructure/base/history_command.cpp
@@ -1,7 +1,7 @@
 #include <readline/history.h>
 #include <readline/readline.h>
 #include "infrastructure/base/command.h"
-#include "infrastructure/utils/log/log.h"
+#include "infrastructure/utils/log-util/log_util.h"
 
 
 XUANSONG_NAMESPACE_HEADER_START
