+++ infrastructure/kernel/CMakeLists.txt
@@ -1,8 +1,8 @@
 aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR} SRC_FILES)
 
-add_library(EquivFusionKernal STATIC ${SRC_FILES})
+add_library(EquivFusionKernel STATIC ${SRC_FILES})
 
-target_link_libraries(EquivFusionKernal
+target_link_libraries(EquivFusionKernel
     PUBLIC
     ${READLINE_LIBRARY}
     EquivFusionUtils
