+++ infrastructure/utils/log-util/log_util.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 #include <cstdlib>
-#include "infrastructure/utils/log/log.h"
+#include "infrastructure/utils/log-util/log_util.h"
 
 XUANSONG_NAMESPACE_HEADER_START
