+++ libs/Tools/EquivMiter/equiv_miter.cpp
@@ -1,5 +1,5 @@
-#include "infrastructure/utils/log/log.h"
-#include "infrastructure/utils/path/path.h"
+#include "infrastructure/utils/log-util/log_util.h"
+#include "infrastructure/utils/path-util/path_util.h"
 #include "libs/Tools/EquivMiter/equiv_miter.h"
 #include "infrastructure/managers/equivfusion_manager/equivfusionManager.h"
